services/api/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
pydantic>=2.0,<3.0
redis>=5.0,<6.0
arq>=0.26.0,<1.0
boto3>=1.35.0,<2.0
python-dotenv>=1.0,<2.0
google-generativeai>=0.8.0,<1.0
PyJWT>=2.8.0,<3.0
google-auth>=2.29.0,<3.0
